animations

+++ app/(tabs)/HistoryScreen.tsx
@@ -5,6 +5,7 @@ import {
   TouchableOpacity,
   ScrollView,
   RefreshControl,
+  Animated,
 } from "react-native";
 import ConversationItem, {
   Conversation,
@@ -17,10 +18,25 @@ import * as constants from "@/constants";
 
 const HistoryScreen: React.FC = () => {
   const [conversations, setConversations] = React.useState<Conversation[]>([]);
+  const [animatedValues, setAnimatedValues] = React.useState<{
+    [key: number]: Animated.Value;
+  }>({});
   const navigation = useNavigation<NavigationProp<TabNavigatorParamList>>();
   const isFocused = navigation.isFocused();
   const [refreshingConversations, setRefreshingConversations] =
     React.useState(false);
+
+  const initializeAnimatedValues = (conversations: Conversation[]) => {
+    const values = conversations.reduce(
+      (acc, conversation) => {
+        acc[conversation.id] = new Animated.Value(1);
+        return acc;
+      },
+      {} as { [key: number]: Animated.Value },
+    );
+    setAnimatedValues(values);
+  };
+
   const fetchConversations = useCallback(async () => {
     setRefreshingConversations(true);
     try {
@@ -34,6 +50,7 @@ const HistoryScreen: React.FC = () => {
         config,
       );
       setConversations(response.data);
+      initializeAnimatedValues(response.data);
     } catch (error: any) {
       alert("Error\n" + error.message);
     }
@@ -41,18 +58,32 @@ const HistoryScreen: React.FC = () => {
   }, []);
 
   const deleteConversation = async (id: number) => {
-    try {
-      const config = {
-        headers: {
-          Authorization: `Bearer ${await AsyncStorage.getItem("token")}`,
-        },
-      };
-      await axios.delete(`${constants.API_URL}/conversation/${id}`, config);
-    } catch (error: any) {
-      alert("Error\n" + error.message);
+    const animatedValue = animatedValues[id];
+    if (animatedValue) {
+      Animated.timing(animatedValue, {
+        toValue: 0,
+        duration: 300,
+        useNativeDriver: false,
+      }).start(async () => {
+        try {
+          const config = {
+            headers: {
+              Authorization: `Bearer ${await AsyncStorage.getItem("token")}`,
+            },
+          };
+          await axios.delete(`${constants.API_URL}/conversation/${id}`, config);
+          setConversations((prevConversations) =>
+            prevConversations.filter((conversation) => conversation.id !== id),
+          );
+          const newAnimatedValues = { ...animatedValues };
+          delete newAnimatedValues[id];
+          setAnimatedValues(newAnimatedValues);
+        } catch (error: any) {
+          alert("Error\n" + error.message);
+        }
+        console.log("Deleted conversation with id", id);
+      });
     }
-    fetchConversations().then();
-    console.log("Deleted conversation with id", id);
   };
 
   React.useEffect(() => {
@@ -77,13 +108,29 @@ const HistoryScreen: React.FC = () => {
           refreshing={refreshingConversations}
           onRefresh={fetchConversations}
         />
-        {conversations.map((item) => (
-          <ConversationItem
-            key={item.id}
-            conversation={item}
-            onTrashPress={() => deleteConversation(item.id)}
-          />
-        ))}
+        {conversations.map((item) => {
+          const animatedValue = animatedValues[item.id];
+          return (
+            <Animated.View
+              key={item.id}
+              style={{
+                opacity: animatedValue,
+                height: animatedValue
+                  ? animatedValue.interpolate({
+                      inputRange: [0, 1],
+                      outputRange: [0, 115],
+                    })
+                  : 60,
+                overflow: "hidden",
+              }}
+            >
+              <ConversationItem
+                conversation={item}
+                onTrashPress={() => deleteConversation(item.id)}
+              />
+            </Animated.View>
+          );
+        })}
       </ScrollView>
     </SafeAreaView>
   );
